Struct atom::AtomSetOnce

pub struct AtomSetOnce<P> where
    P: IntoRawPtr + FromRawPtr,  { /* fields omitted */ }

This is a restricted version of the Atom. It allows for only set_if_none to be called.

swap and take can be used only with a mutable reference. Meaning that AtomSetOnce is not usable as a

Methods

impl<P> AtomSetOnce<P> where
P: IntoRawPtr + FromRawPtr,

pub fn empty() -> AtomSetOnce<P>

Create an empty AtomSetOnce

pub fn new(value: P) -> AtomSetOnce<P>

Create a new AtomSetOnce from Pointer P

pub fn set_if_none(&self, v: P) -> Option<P>

This will do a CAS setting the value only if it is NULL this will return OK(()) if the value was written, otherwise a Err(P) will be returned, where the value was the same value that you passed into this function

pub fn into_atom(self) -> Atom<P>

Convert an AtomSetOnce into an Atom

pub fn atom(&mut self) -> &mut Atom<P>

Allow access to the atom if exclusive access is granted

pub fn is_none(&self) -> bool

Check to see if an atom is None

This only means that the contents was None when it was measured

impl<T, P> AtomSetOnce<P> where
P: IntoRawPtr + FromRawPtr + Deref<Target = T>,

pub fn get<'a>(&'a self) -> Option<&'a T>

If the Atom is set, get the value

impl<T> AtomSetOnce<Box<T>>

pub fn get_mut<'a>(&'a mut self) -> Option<&'a mut T>

If the Atom is set, get the value

impl<T> AtomSetOnce<T> where
T: Clone + IntoRawPtr + FromRawPtr,

pub fn dup<'a>(&self) -> Option<T>

Duplicate the inner pointer if it is set