Crate atom

Structs

Atom
An Atom wraps an AtomicPtr, it allows for safe mutation of an atomic into common Rust Types.

AtomSetOnce
This is a restricted version of the Atom. It allows for only set_if_none to be called.

Traits

FromRawPtr
Convert from a raw ptr into a pointer

GetNextMut
This is a utility Trait that fetches the next ptr from an object.

IntoRawPtr
Convert from into a raw pointer